+++ docs/CHANGELOG.md
@@ -2,7 +2,7 @@ Changelog
 =========
 ### (TBA)
 - Enh: Fix maximum number of allowed accounts
-- Enh: Use maximum number also for space selection
+- Enh: add setting: maximum number of allowed spaces
 
 ### 1.0.0-beta.3 (4/20/2023)
 - Enh: People and spaces page