+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 - Enh: Fix maximum number of allowed accounts
 - Enh: add setting: maximum number of allowed spaces
 - Enh: VerifiedIcon widget options
+- Enh: Send notifications
 
 ### 1.0.0-beta.3 (4/20/2023)
 - Enh: People and spaces page