+++ docs/CHANGELOG.md
@@ -1,11 +1,13 @@
 Changelog
 =========
-### (TBA)
+### 1.0.0 (6/2/2023)
 - Fix: Profile Image upload/edit
 - Enh: Fix maximum number of allowed accounts
 - Enh: add setting: maximum number of allowed spaces
 - Enh: VerifiedIcon widget options
 - Enh: Send notifications
+- Enh: Subscription Payments
+- Chg: `minVersion` to v1.15
 
 ### 1.0.0-beta.3 (4/20/2023)
 - Enh: People and spaces page